+++ src/mysql_shell/executors/errors/runtime.py
@@ -1,11 +1,20 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+import json
+
 
 class ExecutionError(RuntimeError):
     """MySQL shell execution error."""
 
-    # The context attribute is overridden so passwords
-    # are not logged in when showing the traceback
-    # of the previously raised exception
-    __context__ = None
+    def __init__(self, message: str | None = None):
+        """Initialize the error."""
+        if not message:
+            return
+
+        if not message.startswith("MySQL Error"):
+            message = json.loads(message)
+            message = message["error"] if message else message
+            message = message["message"] if message else message
+
+        super().__init__(message)